requirements.txt
Django==1.6.2
oauthlib==0.6.1
python-linkedin==4.0
requests==2.2.1
requests-oauthlib==0.4.0
